added a todo, which i now intend to do, will update the fork as necessary
I Propose, and will now work on fixing it so that it will still gather science which has zero value IF there is a science lab onboard.  Will also try to make it so it checks the lab and only adds the zero value science if the lab does not already have that experiment.